Desktop/projects/fastapi/fastapi-sample-project/requirements.txt


fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
python-multipart==0.0.6
pydantic==2.12.3
psycopg2-binary==2.9.11
alembic==1.12.1
python-dotenv==1.0.0

